Move convert module to shape module
